esm - more AMD removal
