+++ .eslintrc.json
@@ -758,7 +758,6 @@
 						// imports that are allowed in all /test/ files
 						"when": "test",
 						"allow": [
-							"vs/css.build",
 							"vs/css.build.js",
 							"assert",
 							"sinon",
@@ -1010,15 +1009,7 @@
 							"vs/code/~",
 							{
 								"when": "hasBrowser",
-								"pattern": "vs/workbench/workbench.web.main"
-							},
-							{
-								"when": "hasBrowser",
 								"pattern": "vs/workbench/workbench.web.main.js"
-							},
-							{
-								"when": "hasBrowser",
-								"pattern": "vs/workbench/workbench.web.main.internal"
 							},
 							{
 								"when": "hasBrowser",
@@ -1070,13 +1061,11 @@
 							"vs/platform/*/~",
 							"vs/editor/~",
 							"vs/editor/contrib/*/~",
-							"vs/editor/editor.all",
 							"vs/editor/editor.all.js",
 							"vs/workbench/~",
 							"vs/workbench/api/~",
 							"vs/workbench/services/*/~",
 							"vs/workbench/contrib/*/~",
-							"vs/workbench/contrib/terminal/terminal.all",
 							"vs/workbench/contrib/terminal/terminal.all.js"
 						]
 					},
@@ -1089,13 +1078,11 @@
 							"vs/platform/*/~",
 							"vs/editor/~",
 							"vs/editor/contrib/*/~",
-							"vs/editor/editor.all",
 							"vs/editor/editor.all.js",
 							"vs/workbench/~",
 							"vs/workbench/api/~",
 							"vs/workbench/services/*/~",
 							"vs/workbench/contrib/*/~",
-							"vs/workbench/workbench.common.main",
 							"vs/workbench/workbench.common.main.js"
 						]
 					},
@@ -1108,13 +1095,11 @@
 							"vs/platform/*/~",
 							"vs/editor/~",
 							"vs/editor/contrib/*/~",
-							"vs/editor/editor.all",
 							"vs/editor/editor.all.js",
 							"vs/workbench/~",
 							"vs/workbench/api/~",
 							"vs/workbench/services/*/~",
 							"vs/workbench/contrib/*/~",
-							"vs/workbench/workbench.common.main",
 							"vs/workbench/workbench.common.main.js"
 						]
 					},
@@ -1127,13 +1112,11 @@
 							"vs/platform/*/~",
 							"vs/editor/~",
 							"vs/editor/contrib/*/~",
-							"vs/editor/editor.all",
 							"vs/editor/editor.all.js",
 							"vs/workbench/~",
 							"vs/workbench/api/~",
 							"vs/workbench/services/*/~",
 							"vs/workbench/contrib/*/~",
-							"vs/workbench/workbench.common.main",
 							"vs/workbench/workbench.common.main.js"
 						]
 					},
