+++ src/vs/workbench/contrib/performance/browser/perfviewEditor.ts
@@ -17,7 +17,6 @@ import { IExtensionService } from '../../../services/extensions/common/extension
 import { IDisposable, dispose } from '../../../../base/common/lifecycle.js';
 import { ICodeEditorService } from '../../../../editor/browser/services/codeEditorService.js';
 import { writeTransientState } from '../../codeEditor/browser/toggleWordWrap.js';
-import { LoaderStats } from '../../../../base/common/amd.js';
 import { IProductService } from '../../../../platform/product/common/productService.js';
 import { ITextFileService } from '../../../services/textfile/common/textfiles.js';
 import { IEditorService } from '../../../services/editor/common/editorService.js';
@@ -143,11 +142,10 @@ class PerfModelContentProvider implements ITextModelContentProvider {
 		]).then(() => {
 			if (this._model && !this._model.isDisposed()) {
 
-				const stats = LoaderStats.get();
 				const md = new MarkdownBuilder();
 				this._addSummary(md);
 				md.blank();
-				this._addSummaryTable(md, stats);
+				this._addSummaryTable(md);
 				md.blank();
 				this._addExtensionsTable(md);
 				md.blank();
@@ -186,7 +184,7 @@ class PerfModelContentProvider implements ITextModelContentProvider {
 		md.li(`Empty Workspace: ${metrics.emptyWorkbench}`);
 	}
 
-	private _addSummaryTable(md: MarkdownBuilder, stats?: LoaderStats): void {
+	private _addSummaryTable(md: MarkdownBuilder): void {
 
 		const metrics = this._timerService.startupMetrics;
 		const contribTimings = Registry.as<IWorkbenchContributionsRegistry>(WorkbenchExtensions.Workbench).timings;
@@ -200,7 +198,7 @@ class PerfModelContentProvider implements ITextModelContentProvider {
 		table.push(['create window', metrics.timers.ellapsedWindowCreate, '[main]', `initial startup: ${metrics.initialStartup}, ${metrics.initialStartup ? `state: ${metrics.timers.ellapsedWindowRestoreState}ms, widget: ${metrics.timers.ellapsedBrowserWindowCreate}ms, show: ${metrics.timers.ellapsedWindowMaximize}ms` : ''}`]);
 		table.push(['app.isReady => window.loadUrl()', metrics.timers.ellapsedWindowLoad, '[main]', `initial startup: ${metrics.initialStartup}`]);
 		table.push(['window.loadUrl() => begin to import(workbench.desktop.main.js)', metrics.timers.ellapsedWindowLoadToRequire, '[main->renderer]', StartupKindToString(metrics.windowKind)]);
-		table.push(['import(workbench.desktop.main.js)', metrics.timers.ellapsedRequire, '[renderer]', `cached data: ${(metrics.didUseCachedData ? 'YES' : 'NO')}${stats ? `, node_modules took ${stats.nodeRequireTotal}ms` : ''}`]);
+		table.push(['import(workbench.desktop.main.js)', metrics.timers.ellapsedRequire, '[renderer]', `cached data: ${(metrics.didUseCachedData ? 'YES' : 'NO')}`]);
 		table.push(['wait for window config', metrics.timers.ellapsedWaitForWindowConfig, '[renderer]', undefined]);
 		table.push(['init storage (global & workspace)', metrics.timers.ellapsedStorageInit, '[renderer]', undefined]);
 		table.push(['init workspace service', metrics.timers.ellapsedWorkspaceServiceInit, '[renderer]', undefined]);
@@ -336,6 +334,42 @@ class MarkdownBuilder {
 	}
 
 	table(header: string[], rows: Array<Array<{ toString(): string } | undefined>>) {
-		this.value += LoaderStats.toMarkdownTable(header, rows);
+		this.value += this.toMarkdownTable(header, rows);
+	}
+
+	private toMarkdownTable(header: string[], rows: Array<Array<{ toString(): string } | undefined>>): string {
+		let result = '';
+
+		const lengths: number[] = [];
+		header.forEach((cell, ci) => {
+			lengths[ci] = cell.length;
+		});
+		rows.forEach(row => {
+			row.forEach((cell, ci) => {
+				if (typeof cell === 'undefined') {
+					cell = row[ci] = '-';
+				}
+				const len = cell.toString().length;
+				lengths[ci] = Math.max(len, lengths[ci]);
+			});
+		});
+
+		// header
+		header.forEach((cell, ci) => { result += `| ${cell + ' '.repeat(lengths[ci] - cell.toString().length)} `; });
+		result += '|\n';
+		header.forEach((_cell, ci) => { result += `| ${'-'.repeat(lengths[ci])} `; });
+		result += '|\n';
+
+		// cells
+		rows.forEach(row => {
+			row.forEach((cell, ci) => {
+				if (typeof cell !== 'undefined') {
+					result += `| ${cell + ' '.repeat(lengths[ci] - cell.toString().length)} `;
+				}
+			});
+			result += '|\n';
+		});
+
+		return result;
 	}
 }
